requirements.txt
bokeh==2.0.1
numpy==1.18.2
pandas==1.0.3
psycopg2==2.8.4
python-decouple==3.3
SQLAlchemy==1.3.15
scikit-learn==0.22.2.post1
selenium